+++ NEWS.md
@@ -1,6 +1,7 @@
 # mlr3misc 0.1.1
 
 * Fixed warnings about partial argument matching.
+* `invoke()`: New argument `.opts` to temporarily set options.
 
 # mlr3misc 0.1.0